requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
qdrant-client==1.14.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pyyaml>=6.0.1
werkzeug==3.0.1
tqdm>=4.66.1 